Rename folder to domain-suggestions
