+++ packages/data/src/domain-suggestions/types.ts
